Update README.md
This repo is a total joke. This plugin doesn't support webRTC, it supports the proprietary approach of Ant Media to piggy-back on webRTC. It's completely unusable unless you use the paid proprietary stuff of Ant-Media.
"We are all very excited to share this plugin with you and expect the community to benefit from the open-source approach we took." (Hamit Demir, employee of Ant Media)
Hahaha. It has nothing to do about Open Source. Ant-Media is just the typical company that not only piggies back on top of Open-Source but also draps it under the fake flag "we are supporting Open Source".
And obviously, the issues have been removed.